styling the about page
